cc_pbl/DeeSeekR1inAzureFoundry/requirements.txt
install azure-ai-inference
azure-core 
equests 
python-dotenv